[hlf-couchdb] switch to upstream couchdb image
Base images are deprecated (see https://github.com/hyperledger/fabric-baseimage),
and it is now advised to rely on official upstream images.
This commit bumps the chart and update couchdb image to the official one
in version 3.1.1. It also change the env var used to define couchdb user
as without it authentication was completely disabled (anonymous admin).